Clarify installation instructions
Not all users are familiar with `yay`. Add text to make it clearer that
Neovim and hledger can be installed using whatever package manager that
the user prefers.

Add text to instruct the user to install and configure nvim-cmp.

Clarify that `cmp-hledger` needs to be installed with a package manager
for Neovim (not a package manager for the operating system). Clarify
that the example is for packer.

Clarify that `nvim-cmp` needs to be configured to accept `cmp-hledger`.

Change the configuration example to include `cmp-hledger` as a source
only for `.ledger` files. Neovim recognises `*.journal`, `*.ledger` and
`*.ldg` files as `ledger` filetype, so this should work on a brand new
Neovim installation. See this file for proof:

https://github.com/neovim/neovim/blob/master/runtime/lua/vim/filetype.lua